Function rcl::rcl_clock_valid

pub unsafe extern "C" fn rcl_clock_valid(clock: *mut rcl_clock_t) -> bool

Check if the clock has valid values. **

This function returns true if the time source appears to be valid.

It will check that the type is not uninitialized, and that pointers

are not invalid.

Note that if data is uninitialized it may give a false positive.

Attribute | Adherence

—————— | ———––

Allocates Memory | No

Thread-Safe | Yes

Uses Atomics | No

Lock-Free | Yes

\param[in] clock the handle to the clock which is being queried

\return true if the source is believed to be valid, otherwise return false. */