Function rcl::rcl_context_get_domain_id

pub unsafe extern "C" fn rcl_context_get_domain_id(
    context: *mut rcl_context_t, 
    domain_id: *mut usize
) -> rcl_ret_t

Returns the context domain id. **

\pre If context is uninitialized, then it is undefined behavior.

Attribute | Adherence

—————— | ———––

Allocates Memory | No

Thread-Safe | Yes [1]

Uses Atomics | No

Lock-Free | No

[1] Calling the function asynchronously with rcl_init() or rcl_shutdown() can result

in the function sometimes succeeding and sometimes returning #RCL_RET_INVALID_ARGUMENT.

\param[in] context from which the domain id should be retrieved.

\param[out] domain_id output variable where the domain id will be returned.

\return #RCL_RET_INVALID_ARGUMENT if context is invalid (see rcl_context_is_valid()), or

\return #RCL_RET_INVALID_ARGUMENT if domain_id is NULL, or

\return #RCL_RET_OK if the domain id was correctly retrieved. */