Function rcl::rcl_context_get_init_options

pub unsafe extern "C" fn rcl_context_get_init_options(
    context: *const rcl_context_t
) -> *const rcl_init_options_t

Return the init options used during initialization for this context. **

This function can fail and return NULL if:

context is NULL

context is zero-initialized, e.g. context->impl is NULL

If context is uninitialized then that is undefined behavior.

If NULL is returned an error message will have been set.

The options are for reference only, and therefore the returned pointer is

const.

Changing the values in the options is undefined behavior but will likely

have no effect.

Attribute | Adherence

—————— | ———––

Allocates Memory | No

Thread-Safe | Yes

Uses Atomics | Yes

Lock-Free | Yes

\param[in] context object from which the init options should be retrieved

\return pointer to the the init options, or

\return NULL if there was an error */