Function rcl::rcl_get_service_names_and_types_by_node

pub unsafe extern "C" fn rcl_get_service_names_and_types_by_node(
    node: *const rcl_node_t, 
    allocator: *mut rcl_allocator_t, 
    node_name: *const c_char, 
    node_namespace: *const c_char, 
    service_names_and_types: *mut rcl_names_and_types_t
) -> rcl_ret_t

Return a list of service names and types associated with a node. **

The node parameter must point to a valid node.

The service_names_and_types parameter must be allocated and zero initialized.

This function allocates memory for the returned list of names and types and so it is the

callers responsibility to pass service_names_and_types to rcl_names_and_types_fini()

when it is no longer needed.

Failing to do so will result in leaked memory.

\see rcl_get_publisher_names_and_types_by_node for details on the no_demangle parameter.

The returned names are not automatically remapped by this function.

Attempting to create service clients using names returned by this function may not

result in the desired service name being used depending on the remap rules in use.

Attribute | Adherence

—————— | ———––

Allocates Memory | Yes

Thread-Safe | No

Uses Atomics | No

Lock-Free | Maybe [1]

[1] implementation may need to protect the data structure with a lock

\param[in] node the handle to the node being used to query the ROS graph

\param[in] allocator allocator to be used when allocating space for strings

\param[in] node_name the node name of the services to return

\param[in] node_namespace the node namespace of the services to return

\param[out] service_names_and_types list of service names and their types

\return #RCL_RET_OK if the query was successful, or

\return #RCL_RET_NODE_INVALID if the node is invalid, or

\return #RCL_RET_INVALID_ARGUMENT if any arguments are invalid, or

\return #RCL_RET_NODE_INVALID_NAME if the node name is invalid, or

\return #RCL_RET_NODE_INVALID_NAMESPACE if the node namespace is invalid, or

\return #RCL_RET_NODE_NAME_NON_EXISTENT if the node name wasn’t found, or

\return #RCL_RET_ERROR if an unspecified error occurs. */