Function rcl::rcl_init_options_copy

pub unsafe extern "C" fn rcl_init_options_copy(
    src: *const rcl_init_options_t, 
    dst: *mut rcl_init_options_t
) -> rcl_ret_t

Copy the given source init_options to the destination init_options. **

The allocator from the source is used for any allocations and stored in the

destination.

The destination should either be zero initialized with

rcl_get_zero_initialized_init_options() or should have had

rcl_init_options_fini() called on it.

Giving an already initialized init options for the destination will result

in a failure with return code #RCL_RET_ALREADY_INIT.

Attribute | Adherence

—————— | ———––

Allocates Memory | Yes

Thread-Safe | No

Uses Atomics | Yes

Lock-Free | Yes

\param[in] src rcl_init_options_t object to be copied from

\param[out] dst rcl_init_options_t object to be copied into

\return #RCL_RET_OK if the copy is successful, or

\return #RCL_RET_ALREADY_INIT if the dst has already be initialized, or

\return #RCL_RET_INVALID_ARGUMENT if any arguments are invalid, or

\return #RCL_RET_BAD_ALLOC if allocating memory failed, or

\return #RCL_RET_ERROR if an unspecified error occurs. */