Function rcl::rcl_subscription_get_publisher_count

pub unsafe extern "C" fn rcl_subscription_get_publisher_count(
    subscription: *const rcl_subscription_t, 
    publisher_count: *mut usize
) -> rmw_ret_t

Get the number of publishers matched to a subscription. **

Used to get the internal count of publishers matched to a subscription.

Attribute | Adherence

—————— | ———––

Allocates Memory | No

Thread-Safe | Yes

Uses Atomics | Maybe [1]

Lock-Free | Maybe [1]

[1] only if the underlying rmw doesn’t make use of this feature

\param[in] subscription pointer to the rcl subscription

\param[out] publisher_count number of matched publishers

\return #RCL_RET_OK if the count was retrieved, or

\return #RCL_RET_INVALID_ARGUMENT if any arguments are invalid, or

\return #RCL_RET_SUBSCRIPTION_INVALID if the subscription is invalid, or

\return #RCL_RET_ERROR if an unspecified error occurs. */