Function rcl::rcl_timer_reset

pub unsafe extern "C" fn rcl_timer_reset(timer: *mut rcl_timer_t) -> rcl_ret_t

Reset a timer. **

This function can be called on a timer, canceled or not.

For all timers it will reset the last call time to now.

For canceled timers it will additionally make the timer not canceled.

Attribute | Adherence

—————— | ———––

Allocates Memory | No

Thread-Safe | Yes

Uses Atomics | Yes

Lock-Free | Yes [1]

[1] if atomic_is_lock_free() returns true for atomic_int_least64_t

\param[inout] timer the timer to be reset

\return #RCL_RET_OK if the last call time was retrieved successfully, or

\return #RCL_RET_INVALID_ARGUMENT if any arguments are invalid, or

\return #RCL_RET_TIMER_INVALID if the timer is invalid, or

\return #RCL_RET_ERROR an unspecified error occur. */