Function rcl::rcutils_array_list_remove

pub unsafe extern "C" fn rcutils_array_list_remove(
    array_list: *mut rcutils_array_list_t, 
    index: usize
) -> rcutils_ret_t

Removes an entry in the list at the provided index **

This function removes data from the list at the specified index. The capacity

of the list will never decrease when entries are removed.

Attribute | Adherence

—————— | ———––

Allocates Memory | No

Thread-Safe | No

Uses Atomics | No

Lock-Free | Yes

\param[in] array_list to add the data to

\param[in] index the index of the item to remove from the list

\return #RCUTILS_RET_OK if successful, or

\return #RCUTILS_RET_INVALID_ARGUMENT for invalid arguments, or

\return #RCUTILS_RET_INVALID_ARGUMENT if index out of bounds, or

\return #RCUTILS_RET_ERROR if an unknown error occurs. */