Function rcl::rcutils_char_array_resize

pub unsafe extern "C" fn rcutils_char_array_resize(
    char_array: *mut rcutils_char_array_t, 
    new_size: usize
) -> rcutils_ret_t

Resize the internal buffer of the char array. **

The internal buffer of the char array can be resized dynamically if needed.

If the new size is smaller than the current capacity, then the memory is

truncated.

Be aware, that this will deallocate the memory and therefore invalidates any

pointers to this storage.

If the new size is larger, new memory is getting allocated and the existing

content is copied over.

Note that if the array doesn’t own the current buffer the function just

allocates a new block of memory and copies the contents of the old buffer

instead of resizing the existing buffer.

\param[in] char_array pointer to the instance of rcutils_char_array_t which is being resized

\param[in] new_size the new size of the internal buffer

\return #RCUTILS_RET_OK if successful, or

\return #RCUTILS_RET_INVALID_ARGUMENT if new_size is set to zero

\return #RCUTILS_RET_BAD_ALLOC if memory allocation failed, or

\return #RCUTILS_RET_ERROR if an unexpected error occurs. */