Function rcl::rcutils_logging_format_message

pub unsafe extern "C" fn rcutils_logging_format_message(
    location: *const rcutils_log_location_t, 
    severity: c_int, 
    name: *const c_char, 
    timestamp: rcutils_time_point_value_t, 
    msg: *const c_char, 
    logging_output: *mut rcutils_char_array_t
) -> rcutils_ret_t

Formats a log message according to RCUTILS_CONSOLE_OUTPUT_FORMAT **

A formatter that is meant to be used by an output handler to format a log message to the match

the format specified in RCUTILS_CONSOLE_OUTPUT_FORMAT by performing token replacement.

Attribute | Adherence

—————— | ———––

Allocates Memory | Yes

Thread-Safe | No

Uses Atomics | No

Lock-Free | Yes

\param[in] location The location information about where the log came from

\param[in] severity The severity of the log message expressed as an integer

\param[in] name The name of the logger that this message came from

\param[in] timestamp The time at which the log message was generated

\param[in] msg The message being logged

\param[out] logging_output An output buffer for the formatted message

\return #RCUTILS_RET_OK if successful.

\return #RCUTILS_RET_BAD_ALLOC if memory allocation error occured */