Function rcl::rcutils_logging_severity_level_from_string

pub unsafe extern "C" fn rcutils_logging_severity_level_from_string(
    severity_string: *const c_char, 
    allocator: rcutils_allocator_t, 
    severity: *mut c_int
) -> rcutils_ret_t

Get a severity value from its string representation (e.g. DEBUG). **

String representation must match one of the values in

g_rcutils_log_severity_names, but is not case-sensitive.

Examples: UNSET, DEBUG, INFO, WARN, Error, fatal.

\param[in] severity_string String representation of the severity, must be a

null terminated c string

\param[in] allocator rcutils_allocator_t to be used

\param[in,out] severity The severity level as a represented by the

RCUTILS_LOG_SEVERITY enum

\return #RCUTILS_RET_OK if successful, or

\return #RCUTILS_RET_INVALID_ARGUMENT on invalid arguments, or

\return #RCUTILS_RET_LOGGING_SEVERITY_STRING_INVALID if unable to match

string, or

\return #RCUTILS_RET_ERROR if an unspecified error occured. */