Function rcl::rcutils_string_array_init

pub unsafe extern "C" fn rcutils_string_array_init(
    string_array: *mut rcutils_string_array_t, 
    size: usize, 
    allocator: *const rcutils_allocator_t
) -> rcutils_ret_t

Initialize a string array with a given size. **

This function will initialize a given, zero initialized, string array to

a given size.

Note that putting a string into the array gives owenship to the array.

Example:

rcutils_allocator_t allocator = rcutils_get_default_allocator();

rcutils_string_array_t string_array = rcutils_get_zero_initialized_string_array();

rcutils_ret_t ret = rcutils_string_array_init(&string_array, 2, &allocator);

if (ret != RCUTILS_RET_OK) {

// … error handling

}

string_array.data[0] = rcutils_strdup(“Hello”, &allocator);

string_array.data[1] = rcutils_strdup(“World”, &allocator);

ret = rcutils_string_array_fini(&string_array);

\param[inout] string_array object to be initialized

\param[in] size the size the array should be

\param[in] allocator to be used to allocate and deallocate memory

\return #RCUTILS_RET_OK if successful, or

\return #RCUTILS_RET_INVALID_ARGUMENT for invalid arguments, or

\return #RCUTILS_RET_BAD_ALLOC if memory allocation fails, or

\return #RCUTILS_RET_ERROR if an unknown error occurs. */