Function rcl::rmw_create_guard_condition

pub unsafe extern "C" fn rmw_create_guard_condition(
    context: *mut rmw_context_t
) -> *mut rmw_guard_condition_t

Create a guard condition and return a handle to that guard condition. **

This function can fail, and therefore return NULL, if:

context is NULL

context is invalid

memory allocation fails during guard condition creation

an unspecified error occurs

The context must be non-null and valid, i.e. it has been initialized

by rmw_init() and has not been finalized by rmw_shutdown().

Attribute | Adherence

—————— | ———––

Allocates Memory | Yes

Thread-Safe | No

Uses Atomics | No [1]

Lock-Free | No [1]

[1] rmw implementation defined, check the implementation documentation

This should be defined by the rmw implementation.

\param[in] context init context that this node should be associated with

\return rmw guard condition handle or NULL if there was an error */