Function rcl::rmw_get_serialized_message_size

pub unsafe extern "C" fn rmw_get_serialized_message_size(
    type_support: *const rosidl_message_type_support_t, 
    message_bounds: *const rosidl_runtime_c__Sequence__bound, 
    size: *mut usize
) -> rmw_ret_t

Compute the size of a serialized message. **

Given a message definition and bounds, compute the serialized size.

\param[in] type_support The type support of the message to compute.

\param[in] bounds Artifical bounds to use on unbounded fields.

\param[out] size The computed size of the serialized message.

\return RMW_RET_OK if successful, or

\return RMW_RET_INVALID_ARGUMENT if either argument is null, or

\return RMW_RET_UNSUPPORTED if it’s unimplemented, or

\return RMW_RET_ERROR if an unexpected error occurs. */