Function rcl::rmw_subscription_get_actual_qos

pub unsafe extern "C" fn rmw_subscription_get_actual_qos(
    subscription: *const rmw_subscription_t, 
    qos: *mut rmw_qos_profile_t
) -> rmw_ret_t

Retrieve the actual qos settings of the subscription. **

Query the underlying middleware to determine the qos settings

of the subscription.

The actual configuration applied when using RMW_*_SYSTEM_DEFAULT

can only be resolved after the creation of the subscription, and it

depends on the underlying rmw implementation.

If the underlying setting in use can’t be represented in ROS terms,

it will be set to RMW_*_UNKNOWN.

\note The value of avoid_ros_namespace_conventions field is not resolved

with this function. The rcl function rcl_subscription_get_actual_qos()

resolves it.

Attribute | Adherence

—————— | ———––

Allocates Memory | Maybe [1]

Thread-Safe | No

Uses Atomics | Maybe [1]

Lock-Free | Maybe [1]

[1] rmw implementation defined, check the implementation documentation

\param[in] subscription the subscription object to inspect

\param[out] qos the actual qos settings

\return RMW_RET_OK if successful, or

\return RMW_RET_INVALID_ARGUMENT if either argument is null, or

\return RMW_RET_INCORRECT_RMW_IMPLEMENTATION if subscription

implementation identifier does not match, or

\return RMW_RET_ERROR if an unexpected error occurs. */