Function rcl::rmw_topic_endpoint_info_set_topic_type

pub unsafe extern "C" fn rmw_topic_endpoint_info_set_topic_type(
    topic_endpoint_info: *mut rmw_topic_endpoint_info_t, 
    topic_type: *const c_char, 
    allocator: *mut rcutils_allocator_t
) -> rmw_ret_t

Set the topic type in the given topic endpoint info data structure. **

This functions allocates memory and copies the value of the topic_type

argument to set the data structure topic_type member.

Attribute | Adherence

—————— | ———––

Allocates Memory | Yes

Thread-Safe | No

Uses Atomics | No

Lock-Free | Yes

\par Thread-safety

Setting a member is a reentrant procedure, but:

Access to the topic endpoint info data structure is not synchronized.

It is not safe to read or write the `topic_type` member of the given `topic_endpoint`

while setting it.

Access to C-style string arguments is read-only but it is not synchronized.

Concurrent `topic_type` reads are safe, but concurrent reads and writes are not.

The default allocators are thread-safe objects, but any custom allocator may not be.

Check your allocator documentation for further reference.

\pre Given topic_type is a valid C-style string i.e. NULL terminated.

\param[inout] topic_endpoint_info Data structure to be populated.

\param[in] topic_type Type name to be set.

\param[in] allocator Allocator to be used.

\returns RMW_RET_OK if successful, or

\returns RMW_RET_INVALID_ARGUMENT if topic_endpoint_info is NULL, or

\returns RMW_RET_INVALID_ARGUMENT if topic_type is NULL, or

\returns RMW_RET_BAD_ALLOC if memory allocation fails, or

\returns RMW_RET_ERROR when an unspecified error occurs.

\remark This function sets the RMW error state on failure. */